backend/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "Backed for TODO app"
authors = ["Boris Tvaroska <boris@tvaroska.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
google-auth-oauthlib = "^1.2.1"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
python-multipart = "^0.0.17"
jwt = "^1.3.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
